Pull layer features organization into a custom hook
